Update integer fields to use i32 not i64 (#67)
#### Why do we need this change?

The GraphQL spec states that integers should be i32.  I mistakenly made them
i64.

#### What effects does this change have?

Updates cynic to use i32 for integer fields, in line with the spec.

Fixes #52

+++ cynic/src/scalar.rs
@@ -17,7 +17,7 @@ where
     })
 }
 
-impl Scalar for i64 {
+impl Scalar for i32 {
     fn decode(value: &serde_json::Value) -> Result<Self, DecodeError> {
         json_decode::integer().decode(value)
     }
